+++ CHANGELOG.md
@@ -8,6 +8,14 @@ To publish a new release run `scripts/release` from the project directory.
 
 ## [Unreleased]
 
+### Changed
+
+- Replace `ureq` with `reqwest` (see [GH-229]).
+    This implies that the default build now creates a binary linked against the system standard SSL library, i.e. openssl under Linux.
+    A fully static build now requires `--no-default-features --features static` for `cargo build`.
+
+[GH-229]: https://github.com/swsnr/mdcat/pull/229
+
 ## [0.30.3] – 2022-12-01
 
 ### Fixed